# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TDS GEO — AI SEO Automation API",
    "version": "2.0.0",
    "description": "Multi-CMS content generation and SEO optimization platform for Shopify and WordPress stores.",
    "contact": {
      "name": "Traffic Digital Solutions",
      "email": "web.development@trafficdigitalsolutions.com"
    }
  },
  "servers": [
    { "url": "https://16.192.29.174.nip.io", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/clients": {
      "get": {
        "summary": "List clients",
        "tags": ["Clients"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Clients list" }
        }
      },
      "post": {
        "summary": "Create a client",
        "tags": ["Clients"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Client created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}